separate Chart into multiple components
