+++ plugins/performance-resources/src/components/dashboard/Chart.svelte
@@ -1,135 +1,146 @@
 <script lang="ts">
   import { onMount, afterUpdate, createEventDispatcher, onDestroy } from 'svelte'
-  import { Chart, ChartOptions, registerables } from 'chart.js'
   import { EmployeeKRA, ReviewSession, WithKRA, type KRA } from '@hcengineering/performance'
-  import contact, { getName, Person, PersonAccount } from '@hcengineering/contact'
-  import { getClient } from '@hcengineering/presentation'
+  import contact, { getName, PersonAccount } from '@hcengineering/contact'
+  import { getClient, createQuery } from '@hcengineering/presentation'
   import performance from '../../plugin'
   import { Ref, Timestamp } from '@hcengineering/core'
   import { calculateCompletionLevel } from '../../utils/kra'
-  import { Button, DatePresenter, themeStore } from '@hcengineering/ui'
+  import { Button, DatePresenter, Loading, themeStore } from '@hcengineering/ui'
   import kraTeam, { Member } from '@hcengineering/kra-team'
-  import { personIdByAccountId } from '@hcengineering/contact-resources'
-
-  // Register all Chart.js components
-  Chart.register(...registerables)
+  import { personByIdStore, personIdByAccountId } from '@hcengineering/contact-resources'
+  import ChartComponent from './ChartComponent.svelte'
 
   export let space: Ref<ReviewSession>
 
   type KRAsByEmployee = Record<Ref<Member>, Array<KRA & { weight: number, completionLevel: number }>>
 
-  let reviewSession: ReviewSession
-  let employees: Member[] = []
-  let tasks: Array<WithKRA>
-  const taskCompletion: Record<Ref<WithKRA>, number> = {}
-  const kras: KRA[] = []
-  let kraRefs: Ref<KRA>[]
+  let tasks: Array<WithKRA> | undefined = undefined
+  let taskCompletion: Record<Ref<WithKRA>, number> | undefined = undefined
+  let kras: KRA[] | undefined = undefined
+  let kraRefs: Ref<KRA>[] | undefined = undefined
   let employeeKras: EmployeeKRA[] = []
-  let chartCanvas: HTMLCanvasElement
-  let chart: Chart
-  let krasByEmployee: KRAsByEmployee
-  const colors: string[] = ['#4285F4', '#EA4335', '#FBBC05', '#34A853', '#8B44AC', '#00ACC1', '#FF7043']
+  let employees: Member[] = []
+
   let startDate: Timestamp | undefined
   let endDate: Timestamp | undefined
 
-  themeStore.subscribe(() => {
-    updateChart()
-  })
-
   const client = getClient()
-  const dispatch = createEventDispatcher()
+  const hierarchy = client.getHierarchy()
+  const reviewSessionQ = createQuery()
+  const employeeKraQ = createQuery()
+  const taskQ = createQuery()
 
-  $: void client
-    .findOne(performance.class.ReviewSession, {
+  $: reviewSessionQ.query(
+    performance.class.ReviewSession,
+    {
       _id: space
-    })
-    .then((result) => {
-      if (result !== undefined) {
-        reviewSession = result
+    },
+    (res) => {
+      if (res.length === 1) {
+        employees = extractMember(res[0])
       }
-    })
-
-  $: void client
-    .findAll(
-      performance.class.EmployeeKRA,
-      {
-        '$lookup.kra.space': space
+    },
+    {
+      lookup: {
+        members: contact.class.PersonAccount
       },
-      {
-        lookup: {
-          kra: performance.class.KRA
-        }
-      }
-    )
-    .then((result) => {
-      if (result !== undefined) {
-        employeeKras = result
-        result.forEach((entry) => {
-          if (entry.$lookup?.kra !== undefined && !kras.some((value) => value._id === entry.$lookup?.kra?._id)) {
-            kras.push(entry.$lookup.kra)
+      limit: 1
+    }
+  )
+
+  $: employeeKraQ.query(
+    performance.class.EmployeeKRA,
+    {
+      '$lookup.kra.space': space
+    },
+    (res) => {
+      if (res !== undefined && res.length > 0) {
+        employeeKras = res
+        const newKras: KRA[] = []
+        res.forEach((entry) => {
+          if (entry.$lookup?.kra !== undefined && !newKras.some((value) => value._id === entry.$lookup?.kra?._id)) {
+            newKras.push(entry.$lookup.kra)
           }
         })
+        kras = newKras
         kraRefs = kras.map((kra) => kra._id)
       }
-    })
-
-  $: if (reviewSession !== undefined) {
-    const persons = reviewSession.members
-      .map((mem) => $personIdByAccountId.get(mem as Ref<PersonAccount>))
-      .filter((m) => m !== undefined) as Ref<Person>[]
-    void client
-      .findAll(kraTeam.mixin.Member, {
-        _id: {
-          $in: persons
-        }
-      })
-      .then((result) => {
-        if (result !== undefined) {
-          employees = result
-        }
-      })
-  }
+    },
+    {
+      lookup: {
+        kra: performance.class.KRA
+      }
+    }
+  )
 
   $: if (kraRefs !== undefined) {
-    void client
-      .findAll(performance.mixin.WithKRA, {
+    taskQ.query(
+      performance.mixin.WithKRA,
+      {
         kra: { $in: kraRefs },
         createdOn: {
           $gte: startDate ?? Number.MIN_SAFE_INTEGER,
           $lte: endDate !== undefined ? endDate + 86400 : Number.MAX_SAFE_INTEGER
         }
-      })
-      .then((result) => {
-        if (result !== undefined) {
-          tasks = result
-        }
-      })
-  }
-
-  $: {
-    if (tasks !== undefined) {
-      const updateCompletionLevels = async (): Promise<void> => {
-        for (const task of tasks) {
-          taskCompletion[task._id] = (await calculateCompletionLevel(task._id)) ?? 0
+      },
+      (res) => {
+        if (res !== undefined) {
+          tasks = res
         }
       }
-      void updateCompletionLevels()
-    }
+    )
+  } else {
+    taskQ.unsubscribe()
   }
 
-  $: {
-    // Group KRAs by employee using the relationships table
-    krasByEmployee = employees.reduce<KRAsByEmployee>((acc, employee) => {
+  $: if (tasks !== undefined) {
+    const updateCompletionLevels = async (): Promise<void> => {
+      const completion: Record<Ref<WithKRA>, number> = {}
+      for (const task of tasks ?? []) {
+        completion[task._id] = (await calculateCompletionLevel(task._id)) ?? 0
+      }
+      taskCompletion = completion
+    }
+    void updateCompletionLevels()
+  }
+
+  function resetFilters (): void {
+    startDate = undefined
+    endDate = undefined
+  }
+
+  function extractMember (reviewSession: ReviewSession): Member[] {
+    return reviewSession.members
+      .map((mem) => $personIdByAccountId.get(mem as Ref<PersonAccount>))
+      .filter((m) => m !== undefined)
+      .map((m) => {
+        const p = $personByIdStore.get(m)
+        if (p !== undefined) {
+          return hierarchy.as(p, kraTeam.mixin.Member)
+        }
+        return undefined
+      })
+      .filter((m) => m !== undefined) as Member[]
+  }
+
+  function summarizeKra (
+    employees: Member[],
+    kras: KRA[],
+    kraAssigns: EmployeeKRA[],
+    tasks: WithKRA[],
+    taskCompletion: Record<Ref<WithKRA>, number>
+  ): KRAsByEmployee {
+    return employees.reduce<KRAsByEmployee>((acc, employee) => {
       // Get all EmployeeKRA entries for this employee
-      const employeeKraEntries = employeeKras.filter((entry) => entry.assignee === employee._id)
+      const employeeKraEntries = kraAssigns.filter((entry) => entry.assignee === employee._id)
       // For each entry, fetch the corresponding KRA details
       const employeeKraDetails = employeeKraEntries
         .map((entry) => {
           const kra = kras.find((k) => k._id === entry.kra)
-          if (kra == null || tasks === undefined) return null
+          if (kra == null) return null
           const filteredTasks = tasks.filter((task) => {
-            const asMixin = client.getHierarchy().as(task, performance.mixin.WithKRA) as any
-            return asMixin.kra === kra._id && asMixin.assignee === employee
+            return task.kra === kra._id && task.assignee === employee._id
           })
           let completionLevel = filteredTasks.reduce<number>((acc, task) => {
             return acc + (taskCompletion[task._id] ?? 0)
@@ -148,206 +159,6 @@
       acc[employee._id] = employeeKraDetails as Array<KRA & { weight: number, completionLevel: number }>
       return acc
     }, {})
-
-    // If data changes and chart exists, update it
-    if (chart !== undefined) {
-      updateChart()
-    }
-  }
-
-  // Create datasets for Chart.js
-  function createChartData (): any {
-    const employeeNames = employees.map(
-      (e) => getName(client.getHierarchy(), e)
-    )
-
-    // Create a dataset for each KRA
-    const datasets = kras.map((kra, index) => {
-      const data = employees.map((employee) => {
-        const employeeKraDetails = krasByEmployee[employee._id]
-        if (employeeKraDetails == null) {
-          return 0
-        }
-        const matchingKra = employeeKraDetails.find((k) => k._id === kra._id)
-
-        // Calculate contribution to performance score
-        const d =
-          matchingKra?.completionLevel != null && matchingKra.weight != null
-            ? matchingKra.completionLevel * matchingKra.weight
-            : 0
-        return d * 100
-      })
-
-      return {
-        label: kra.title,
-        data,
-        backgroundColor: colors[index % colors.length],
-        borderColor: $themeStore.dark ? '#3c3f44' : '#dfe1e4',
-        borderWidth: 1,
-        barPercentage: 0.8
-      }
-    })
-
-    // // Create dataset for remaining percentage to reach 100%
-    // const remainingData = employees.map(employee => {
-    //   return Math.max(0, 100 - employee.performanceScore)
-    // })
-
-    // datasets.push({
-    //   label: 'Remaining',
-    //   data: remainingData,
-    //   backgroundColor: '#E0E0E0', // Light gray for remaining portion
-    //   borderColor: 'white',
-    //   borderWidth: 1,
-    //   barPercentage: 0.8
-    // })
-
-    return {
-      labels: employeeNames,
-      datasets
-    }
-  }
-
-  function createChart (): void {
-    if (chartCanvas == null) return
-
-    // Destroy existing chart if it exists
-    if (chart != null) {
-      chart.destroy()
-    }
-
-    const ctx = chartCanvas.getContext('2d')
-    if (ctx == null) return
-
-    chart = new Chart(ctx, {
-      type: 'bar',
-      data: createChartData(),
-      options: getChartOptions()
-    })
-
-    // Add this to make segments clickable with cursor pointer
-    chartCanvas.style.cursor = 'pointer'
-  }
-
-  function getChartOptions (): ChartOptions {
-    return {
-      indexAxis: 'y', // Horizontal bar chart
-      responsive: true,
-      maintainAspectRatio: false,
-      color: $themeStore.dark ? 'rgba(255, 255, 255, .8)' : '#rgba(0, 0, 0, .8)',
-      scales: {
-        x: {
-          stacked: true,
-          min: 0,
-          max: 100,
-          title: {
-            color: $themeStore.dark ? 'rgba(255, 255, 255, .8)' : '#rgba(0, 0, 0, .8)',
-            display: true,
-            text: 'Performance Score (%)'
-          },
-          grid: {
-            display: true,
-            color: $themeStore.dark ? 'rgba(255, 255, 255, 0.05)' : 'rgba(0, 0, 0, 0.05)'
-          }
-        },
-        y: {
-          stacked: true,
-          title: {
-            color: $themeStore.dark ? 'rgba(255, 255, 255, .8)' : '#rgba(0, 0, 0, .8)',
-            display: true,
-            text: 'Employees'
-          }
-        }
-      },
-      plugins: {
-        legend: {
-          position: 'bottom'
-        },
-        tooltip: {
-          callbacks: {
-            title: (tooltipItems) => {
-              return tooltipItems[0].label // Employee name
-            },
-            footer: (tooltipItems) => {
-              const employeeIndex = tooltipItems[0].dataIndex
-              const employee = employees[employeeIndex]
-              if (employee == null) return ''
-
-              // Show KRA breakdown
-              const employeeKras = krasByEmployee[employee._id]
-              if (employeeKras.length === 0) return ''
-
-              let footerText = '\nKRA Breakdown:'
-              employeeKras.forEach((kra) => {
-                footerText += `\n${kra.title} (${kra.weight * 100}%): ${kra.completionLevel ?? 0}`
-              })
-
-              return footerText
-            }
-          }
-        }
-      },
-      onClick: (event, elements, chart) => {
-        if (elements === undefined || elements.length === 0) return
-
-        const element = elements[0]
-        const index = element.index
-        const datasetIndex = element.datasetIndex
-        const employee = employees[index]
-
-        // Get the clicked KRA data
-        const datasets = chart.data.datasets
-        const dataset = datasets[datasetIndex]
-
-        const kraName = dataset.label
-        const employeeKras = krasByEmployee[employee._id]
-        const kra = employeeKras.find((k) => k.title === kraName)
-
-        if (kra != null) {
-          dispatch('segmentClick', {
-            employee,
-            kra: kra as KRA
-          })
-        }
-      }
-    }
-  }
-
-  function updateChart (): void {
-    if (chart != null) {
-      chart.data = createChartData()
-      chart.options = getChartOptions()
-      chart.update()
-    }
-  }
-
-  onMount(() => {
-    // When component is first mounted, create chart if data is available
-    if (chartCanvas != null && employees.length > 0 && kras.length > 0) {
-      createChart()
-    }
-  })
-
-  afterUpdate(() => {
-    // Check if we need to create or update the chart
-    if (chartCanvas != null && employees.length > 0 && kras.length > 0) {
-      if (chart == null) {
-        createChart()
-      } else {
-        updateChart()
-      }
-    }
-  })
-
-  onDestroy(() => {
-    if (chart != null) {
-      chart.destroy()
-    }
-  })
-
-  function resetFilters (): void {
-    startDate = undefined
-    endDate = undefined
   }
 </script>
 
@@ -378,7 +189,18 @@
       />
       <Button label={performance.string.ResetFilters} kind={'regular'} size={'large'} on:click={resetFilters} />
     </div>
-    <canvas id="chart" bind:this={chartCanvas}></canvas>
+    {#if kras !== undefined && employees !== undefined && employeeKras !== undefined && tasks !== undefined && taskCompletion !== undefined}
+      {@const krasByEmployee = summarizeKra(
+        employees,
+        kras,
+        employeeKras,
+        tasks,
+        taskCompletion
+      )}
+      <ChartComponent {employees} {krasByEmployee} {kras} on:segmentClick />
+    {:else}
+      <Loading />
+    {/if}
   {/if}
 </div>
